refactor: enhance cursor handling and selection checks in editor components
