+++ packages/super-editor/src/components/slash-menu/utils.js
@@ -67,24 +67,6 @@ export const getPropsByItemId = (itemId, props) => {
 }
 
 /**
- * Move the editor cursor to the position closest to the mouse event
- * @param {MouseEvent} event
- * @param {Object} editor - The editor instance
- */
-export function moveCursorToMouseEvent(event, editor) {
-  const { view } = editor;
-  const coords = { left: event.clientX, top: event.clientY };
-  const pos = view.posAtCoords(coords)?.pos;
-  if (typeof pos === 'number') {
-    const tr = view.state.tr.setSelection(
-      TextSelection.create(view.state.doc, pos)
-    );
-    view.dispatch(tr);
-    view.focus();
-  }
-}
-
-/**
  * Get the current editor context for menu logic
  *
  * @param {Object} editor - The editor instance
@@ -121,21 +103,4 @@ export async function getEditorContext(editor, event) {
     event,
     clipboardContent,
   };
-}
-
-/**
- * Checks if the current selection is inside a table node
- * Used to determine whether to show table-specific menu items in the slash menu
- * 
- * @param {EditorState} state - The current editor state
- * @returns {boolean} True if selection is inside a table, false otherwise
- */
-export function isInsideTable(state) {
-  const $from = state.selection.$from;
-  for (let d = $from.depth; d > 0; d--) {
-    if ($from.node(d).type.name === 'table') {
-      return true;
-    }
-  }
-  return false;
 }
